requirements.txt
fastapi
uvicorn[standard]
pydantic-ai
httpx
pandas
PyPDF2
python-docx